quantum: Consolidating quantum kinds into one file (kind.py)
